update version 78

+++ android/src/main/java/com/truvideoreactturbocamerasdk/ArCameraActivity.kt
@@ -94,12 +94,12 @@ class ArCameraActivity : AppCompatActivity() {
             mode = TruvideoSdkCameraMode.videoAndImage(
               imageMaxCount = imageLimit?.toInt(),
               videoMaxCount = videoLimit?.toInt(),
-              durationLimit = videoDurationLimit?.toDouble()
+              durationLimit = videoDurationLimit?.toInt()
             )
           }else if(mediaLimit != null){
             mode = TruvideoSdkCameraMode.videoAndImage(
               maxCount = mediaLimit.toInt(),
-              durationLimit = videoDurationLimit?.toDouble()
+              durationLimit = videoDurationLimit?.toInt()
             )
           }else {
             mode = TruvideoSdkCameraMode.videoAndImage()
@@ -108,7 +108,7 @@ class ArCameraActivity : AppCompatActivity() {
         "video" -> {
           mode = TruvideoSdkCameraMode.video(
             maxCount = videoLimit?.toInt(),
-            durationLimit = videoDurationLimit?.toDouble()
+            durationLimit = videoDurationLimit?.toInt()
           )
         }
         "image" -> {
@@ -121,12 +121,12 @@ class ArCameraActivity : AppCompatActivity() {
         }
         "singleVideo" ->{
           mode = TruvideoSdkCameraMode.singleVideo(
-            durationLimit = videoDurationLimit?.toDouble()
+            durationLimit = videoDurationLimit?.toInt()
           )
         }
         "singleVideoOrImage" -> {
           mode = TruvideoSdkCameraMode.singleVideoOrImage(
-            durationLimit = videoDurationLimit?.toDouble()
+            durationLimit = videoDurationLimit?.toInt()
           )
         }
       }
